backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.1
twilio==8.10.0
bcrypt==4.1.2
flask-pymongo==2.3.0
email-validator==2.1.0.post1
phonenumbers
requests
google-generativeai